model_gen/requirements.txt
fastapi
uvicorn
requests
python-dotenv
python-multipart  